Restructure dashboard into feature folders
