+++ src/main.tsx
@@ -1,15 +1,15 @@
 import { createRoot } from "react-dom/client";
-import ErrorPage from "@pages/error.tsx";
 import { ErrorBoundary } from "react-error-boundary";
 import { Routes, Route, HashRouter, Navigate } from "react-router";
 import Popup from "@pages/popup/popup";
 import Options from "@pages/options/options";
-import Dashboard from "@pages/dashboard/dashboard";
-import { UsageScopePage } from "@pages/dashboard/usage/usageScopePage";
+import Dashboard from "@pages/dashboard/overview/dashboard";
+import { ScopePage } from "@pages/dashboard/scope/scopePage";
 import PlatformUsage from "@pages/dashboard/platform/platformUsage";
-import DashboardErrorPage from "@pages/dashboard/dashboardErrorPage";
-import DashboardNotFound from "@pages/dashboard/dashboardNotFound";
-import OptionsErrorPage from "@pages/options/optionsErrorPage";
+import DashboardErrorPage from "@pages/dashboard/shared/dashboardError";
+import PopupErrorPage from "@pages/popup/popupError";
+import DashboardNotFound from "@pages/dashboard/shared/notFound";
+import OptionsErrorPage from "@pages/options/optionsError";
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 
 import "@styles/global.css";
@@ -18,7 +18,7 @@ import { StrictMode } from "react";
 import { PopupLayout } from "@layouts/popupLayout";
 import DashboardLayout from "@layouts/dashboardLayout";
 import { OptionsLayout } from "@layouts/optionsLayout";
-import { WebsiteUsage } from "@pages/dashboard/websiteUsage";
+import { SiteDetailPage } from "@pages/dashboard/scope/siteDetailPage";
 
 const domRoot = document.querySelector("#root") as HTMLElement;
 
@@ -34,7 +34,7 @@ root.render(
                         <Route
                             index
                             element={
-                                <ErrorBoundary FallbackComponent={ErrorPage}>
+                                <ErrorBoundary FallbackComponent={PopupErrorPage}>
                                     <Popup />
                                 </ErrorBoundary>
                             }
@@ -63,9 +63,9 @@ root.render(
                         <Route index element={<Navigate to="daily" replace />} />
                         <Route path="daily" element={<Dashboard chart={"daily"} />} />
                         <Route path="sites" element={<Dashboard chart={"sites"} />} />
-                        <Route path=":date" element={<UsageScopePage />} />
+                        <Route path=":date" element={<ScopePage />} />
                         <Route path="platform/:platformId" element={<PlatformUsage />} />
-                        <Route path="site/:siteName" element={<WebsiteUsage />} />
+                        <Route path="site/:siteName" element={<SiteDetailPage />} />
                         <Route path="*" element={<DashboardNotFound />} />
                     </Route>
                 </Routes>
